Added localization functions in DirManipulation object changed settings functions and class created SubMenu for localization

+++ src/main/kotlin/MiddleGrid.kt
@@ -23,7 +23,7 @@ import kotlin.io.path.pathString
 @Composable
 fun MiddleGrid(
     list: List<String>,
-    size: Int,
+    size: Float,
     selectedItemsList: MutableList<String>,
     selectedItem: (String) -> Unit
 ) {
